rename: StudyProvider 위치 변경
