+++ apps/admin/app/students/_components/StudyDropDown.tsx
@@ -5,8 +5,8 @@ import type { StudyListApiResponseDto } from "types/dtos/studyList";
 import DropDown from "wowds-ui/DropDown";
 import DropDownOption from "wowds-ui/DropDownOption";
 
+import { studyAtom } from "../_contexts/StudyProvider";
 import DropDownTrigger from "./DropDownTrigger";
-import { studyAtom } from "./StudyProvider";
 
 const StudyDropDown = ({
   studyList,
