+++ apps/admin/app/students/layout.tsx
@@ -2,7 +2,7 @@ import { css } from "@styled-system/css";
 import { styled } from "@styled-system/jsx";
 import Navbar from "components/Navbar";
 
-import { StudyProvider } from "./_components/StudyProvider";
+import { StudyProvider } from "./_contexts/StudyProvider";
 
 const StudentsLayout = ({
   children,
